alpha-2.0
git-svn-id: svn://localhost/svn/uclinux-dist/trunk@60 3e2a5524-39e8-4cb9-84a9-c49da69cba3a

+++ linux-2.6.x/arch/mips/lib-64/dump_tlb.c
@@ -200,10 +200,12 @@ void dump16(unsigned long *p)
 {
 	int i;
 
-	for (i = 0; i < 8; i++) {
-		printk("*%08lx == %08lx, ", (unsigned long)p, *p);
+	for(i = 0; i < 8; i++) {
+		printk("*%08lx == %08lx, ",
+		       (unsigned long)p, (unsigned long)*p);
 		p++;
-		printk("*%08lx == %08lx\n", (unsigned long)p, *p);
+		printk("*%08lx == %08lx\n",
+		       (unsigned long)p, (unsigned long)*p);
 		p++;
 	}
 }
