+++ linux-2.6.x/mm/mincore.c
@@ -14,6 +14,7 @@
 
 #include <asm/uaccess.h>
 #include <asm/pgtable.h>
+#include <asm/pgalloc.h>
 
 /*
  * Later we can get more picky about what "in core" means precisely.
@@ -123,7 +124,7 @@ asmlinkage long sys_mincore(unsigned long start, size_t len,
 		goto out;
 
 	error = -EFAULT;
-	if (!access_ok(VERIFY_WRITE, vec, len >> PAGE_SHIFT))
+	if (!access_ok(VERIFY_WRITE, (unsigned long) vec, len >> PAGE_SHIFT))
 		goto out;
 
 	error = 0;
